Split BI test to common and Tampere specific parts
